Rename azure_core::auth to azure_core::credentials
