+++ sdk/core/azure_core/src/credentials.rs
