+++ sdk/core/azure_core/src/policies/bearer_token_policy.rs
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 
 use crate::{
-    authentication::TokenCredential,
+    credentials::TokenCredential,
     headers::AUTHORIZATION,
     policies::{Policy, PolicyResult},
     Context, Request,
