+++ sdk/core/azure_core_amqp/src/fe2o3/management.rs
@@ -11,7 +11,7 @@ use crate::{
 };
 
 use async_std::sync::Mutex;
-use azure_core::{authentication::AccessToken, error::Result};
+use azure_core::{credentials::AccessToken, error::Result};
 use fe2o3_amqp_management::operations::ReadResponse;
 use fe2o3_amqp_types::{messaging::ApplicationProperties, primitives::SimpleValue};
 use std::sync::{Arc, OnceLock};
