+++ sdk/core/azure_core_amqp/src/management.rs
@@ -6,7 +6,7 @@ use super::{
     session::AmqpSession,
     value::{AmqpOrderedMap, AmqpValue},
 };
-use azure_core::{authentication::AccessToken, error::Result};
+use azure_core::{credentials::AccessToken, error::Result};
 use std::fmt::Debug;
 
 #[cfg(all(feature = "fe2o3-amqp", not(target_arch = "wasm32")))]
