+++ sdk/eventhubs/azure_messaging_eventhubs/src/consumer/mod.rs
@@ -16,7 +16,7 @@ use super::{
 use async_std::sync::Mutex;
 use async_stream::try_stream;
 use azure_core::{
-    authentication::{AccessToken, TokenCredential},
+    credentials::{AccessToken, TokenCredential},
     error::{Error, Result},
     RetryOptions,
 };
@@ -45,7 +45,7 @@ pub struct ConsumerClient {
     session_instances: Mutex<HashMap<String, Arc<AmqpSession>>>,
     mgmt_client: Mutex<OnceLock<ManagementInstance>>,
     connection: OnceLock<AmqpConnection>,
-    credential: Box<dyn azure_core::authentication::TokenCredential>,
+    credential: Box<dyn azure_core::credentials::TokenCredential>,
     eventhub: String,
     url: String,
     authorization_scopes: Mutex<HashMap<String, AccessToken>>,
