+++ sdk/identity/azure_identity/src/credentials/environment_credentials.rs
@@ -7,7 +7,7 @@ use crate::credentials::{
     ClientSecretCredential, TokenCredentialOptions, WorkloadIdentityCredential,
 };
 use azure_core::{
-    authentication::{AccessToken, TokenCredential},
+    credentials::{AccessToken, TokenCredential},
     error::{Error, ErrorKind},
 };
