+++ sdk/identity/azure_identity/src/credentials/imds_managed_identity_credentials.rs
@@ -3,7 +3,7 @@
 
 use crate::{credentials::cache::TokenCache, TokenCredentialOptions};
 use azure_core::{
-    authentication::{AccessToken, Secret, TokenCredential},
+    credentials::{AccessToken, Secret, TokenCredential},
     error::{http_response_from_body, Error, ErrorKind},
     headers::HeaderName,
     json::from_json,
